api/playground/requirements.txt
fastapi
uvicorn
langchain-core
langchain-community
langchain-openai
openai
sse-starlette
pydantic
python-dotenv
replicate

litellm
chromadb